requirements.txt
eventlet
flask_cors
flask
flask_socketio
huggingface_hub
requests
python-dotenv